+++ Flow/src/utopia/flow/async/Wait.scala
@@ -2,7 +2,8 @@ package utopia.flow.async
 
 import utopia.flow.time.{Now, WaitTarget}
 import utopia.flow.time.TimeExtensions._
-import utopia.flow.util.logging.{Logger, SysErrLogger}
+import utopia.flow.util.logging.SysErrLogger
+import utopia.flow.util.CollectionExtensions._
 
 import scala.concurrent.ExecutionContext
 import scala.util.Try
@@ -14,10 +15,18 @@ object Wait
 	  * @param target Target for this wait
 	  * @param lock Wait lock to use (optional)
 	  * @param exc Implicit execution context
+	  * @return Whether the wait completed without being interrupted through an InterruptedException.
+	  *         If false, the process that follows should likewise take measures to hurry to its completion
+	  *         (use discrection).
 	  */
 	def apply(target: WaitTarget, lock: AnyRef = new AnyRef)(implicit exc: ExecutionContext) = {
-		if (target.isPositive)
-			new Wait(target, lock).run()
+		if (target.isPositive) {
+			val wait = new Wait(target, lock)
+			wait.run()
+			wait.state.isNotBroken
+		}
+		else
+			true
 	}
 }
 
@@ -35,32 +44,29 @@ class Wait(val target: WaitTarget, val lock: AnyRef = new AnyRef,
 	
 	override protected def runOnce() =
 	{
-		var waitCompleted = false
-		target.endTime match {
-			case Some(targetTime) =>
-				lock.synchronized {
-					while (!waitCompleted && targetTime.isInFuture) {
-						val waitDuration = targetTime - Now
-						// Performs the actual wait here (nano precision)
-						// Exceptions are ignored
-						Try {
+		// Catches InterruptedExceptions
+		Try {
+			var waitCompleted = false
+			target.endTime match {
+				case Some(targetTime) =>
+					lock.synchronized {
+						while (!waitCompleted && targetTime.isInFuture) {
+							val waitDuration = targetTime - Now
+							// Performs the actual wait here (nano precision)
 							lock.wait(waitDuration.toMillis, waitDuration.getNano / 1000)
-							// May break on any notify call. stop(), however, breaks regardless
+							// May break on any notify call. Also, stop() always breaks
 							if (target.breaksOnNotify || state.isBroken)
 								waitCompleted = true
 						}
 					}
-				}
-			case None =>
-				lock.synchronized {
-					while (!waitCompleted) {
-						// Waits until notified, exceptions are ignored
-						Try {
-							lock.wait()
-							waitCompleted = true
-						}
-					}
-				}
+				// Waits until notified
+				case None => lock.synchronized { lock.wait() }
+			}
+		}.failure.foreach {
+			// Case: Interrupted during waiting => Treats as if broken through stop()
+			case _: InterruptedException => markAsInterrupted()
+			// Case: Interrupted through another exception => Logs
+			case error => SysErrLogger(error, "Unexpected error during a Wait")
 		}
 	}
 }