+++ Reflection/src/utopia/reflection/container/swing/window/Window.scala
@@ -247,22 +247,21 @@ trait Window[+Content <: Stackable with AwtComponentRelated] extends Stackable w
       */
     def startEventGenerators(actorHandler: ActorHandler)(implicit exc: ExecutionContext) =
     {
-        generatorActivated.runAndSet
-        {
-			// Starts mouse listening
+        if (generatorActivated.set()) {
+            // Starts mouse listening
             val mouseEventGenerator = new MouseEventGenerator(content.component)
             actorHandler += mouseEventGenerator
             mouseEventGenerator.buttonHandler += MouseButtonStateListener() { e =>
                 content.distributeMouseButtonEvent(e); None }
             mouseEventGenerator.moveHandler += MouseMoveListener() { content.distributeMouseMoveEvent(_) }
             mouseEventGenerator.wheelHandler += MouseWheelListener() { content.distributeMouseWheelEvent(_) }
-			
+    
             GlobalMouseEventHandler.registerGenerator(mouseEventGenerator)
-            
-			// Starts key listening
+    
+            // Starts key listening
             val keyStatusListener = new KeyStatusListener()
             GlobalKeyboardEventHandler += keyStatusListener
-            
+    
             // Quits event listening once this window finally closes
             uponCloseAction.setOne(() =>
             {