+++ Disciple/Changes.md
@@ -2,6 +2,8 @@
 
 ## v1.5.3 (in development)
 Supports changes in **Flow** v1.17
+### Other Changes
+- **RequestRateLimiter** now handles thread interruptions, as well as JVM shutdowns, better
 
 ## v1.5.2 - 18.08.2022
 New Build / Supports changes in **Flow** v1.16